+++ src/app/features/applications/application/application-base.component.spec.ts
@@ -1,45 +1,20 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterTestingModule } from '@angular/router/testing';
-import { StoreModule } from '@ngrx/store';
 
-import { CoreModule } from '../../../core/core.module';
-import { MDAppModule } from '../../../core/md.module';
-import { SharedModule } from '../../../shared/shared.module';
-import { appReducers } from '../../../store/reducers.module';
-import { getInitialTestStoreState } from '../../../test-framework/store-test-helper';
 import { ApplicationBaseComponent } from './application-base.component';
-import { ApplicationEnvVarsService } from './build-tab/application-env-vars.service';
-import { ApplicationStateService } from '../../../shared/components/application-state/application-state.service';
 
-describe('ApplicationBaseComponent', () => {
+describe('ApplicationTabBaseComponent', () => {
   let component: ApplicationBaseComponent;
   let fixture: ComponentFixture<ApplicationBaseComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [
-        ApplicationBaseComponent,
-      ],
+      declarations: [ ApplicationBaseComponent ],
       imports: [
-        StoreModule,
-        CoreModule,
-        SharedModule,
-        BrowserAnimationsModule,
-        RouterTestingModule,
-        MDAppModule,
-        StoreModule.forRoot(
-          appReducers
-          , {
-            initialState: getInitialTestStoreState()
-          })
-      ],
-      providers: [
-        ApplicationStateService,
-        ApplicationEnvVarsService
+        RouterTestingModule
       ]
     })
-      .compileComponents();
+    .compileComponents();
   }));
 
   beforeEach(() => {
@@ -48,7 +23,7 @@ describe('ApplicationBaseComponent', () => {
     fixture.detectChanges();
   });
 
-  it('should be created', () => {
+  it('should create', () => {
     expect(component).toBeTruthy();
   });
 });
