+++ src/app/features/applications/application/application-tabs-base/application-tabs-base.component.spec.ts
@@ -1,25 +1,54 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { RouterTestingModule } from '@angular/router/testing';
+import { StoreModule } from '@ngrx/store';
 
-import { ApplicationTabsBaseComponent } from './application-tab-base.component';
+import { CoreModule } from '../../../../core/core.module';
+import { MDAppModule } from '../../../../core/md.module';
+import { SharedModule } from '../../../../shared/shared.module';
+import { appReducers } from '../../../../store/reducers.module';
+import { getInitialTestStoreState } from '../../../../test-framework/store-test-helper';
+import { ApplicationBaseComponent } from './../application-base.component';
+import { ApplicationEnvVarsService } from './../application-tabs-base/tabs/build-tab/application-env-vars.service';
+import { ApplicationStateService } from '../../../../shared/components/application-state/application-state.service';
 
-describe('ApplicationTabBaseComponent', () => {
-  let component: ApplicationTabsBaseComponent;
-  let fixture: ComponentFixture<ApplicationTabsBaseComponent>;
+describe('ApplicationTabsBaseComponent', () => {
+  let component: ApplicationBaseComponent;
+  let fixture: ComponentFixture<ApplicationBaseComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ ApplicationTabsBaseComponent ]
+      declarations: [
+        ApplicationBaseComponent,
+      ],
+      imports: [
+        StoreModule,
+        CoreModule,
+        SharedModule,
+        BrowserAnimationsModule,
+        RouterTestingModule,
+        MDAppModule,
+        StoreModule.forRoot(
+          appReducers
+          , {
+            initialState: getInitialTestStoreState()
+          })
+      ],
+      providers: [
+        ApplicationStateService,
+        ApplicationEnvVarsService
+      ]
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(ApplicationTabsBaseComponent);
+    fixture = TestBed.createComponent(ApplicationBaseComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
 
-  it('should create', () => {
+  it('should be created', () => {
     expect(component).toBeTruthy();
   });
 });
