+++ src/app/features/applications/application/application-tabs-base/tabs/events-tab/events-tab.component.spec.ts
@@ -1,17 +1,17 @@
 import { init } from 'protractor/built/launcher';
-import { CoreModule } from '../../../../core/core.module';
-import { MDAppModule } from '../../../../core/md.module';
+import { CoreModule } from '../../../../../../core/core.module';
+import { MDAppModule } from '../../../../../../core/md.module';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { EventsTabComponent } from './events-tab.component';
-import { SharedModule } from '../../../../shared/shared.module';
+import { SharedModule } from '../../../../../../shared/shared.module';
 import { StoreModule } from '@ngrx/store';
-import { appReducers } from '../../../../store/reducers.module';
-import { getInitialTestStoreState } from '../../../../test-framework/store-test-helper';
-import { ApplicationService } from '../../application.service';
-import { ApplicationStateService } from '../../../../shared/components/application-state/application-state.service';
+import { appReducers } from '../../../../../../store/reducers.module';
+import { getInitialTestStoreState } from '../../../../../../test-framework/store-test-helper';
+import { ApplicationService } from '../../../../application.service';
+import { ApplicationStateService } from '../../../../../../shared/components/application-state/application-state.service';
 import { ApplicationEnvVarsService } from '../build-tab/application-env-vars.service';
-import { ApplicationsModule } from '../../applications.module';
+import { ApplicationsModule } from '../../../../applications.module';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
