+++ src/app/features/applications/application/application-tabs-base/tabs/log-stream-tab/log-stream-tab.component.spec.ts
@@ -2,17 +2,17 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterTestingModule } from '@angular/router/testing';
 import { StoreModule } from '@ngrx/store';
 
-import { CoreModule } from '../../../../core/core.module';
-import { MDAppModule } from '../../../../core/md.module';
-import { LogViewerComponent } from '../../../../shared/components/log-viewer/log-viewer.component';
-import { ApplicationSchema, GetApplication } from '../../../../store/actions/application.actions';
-import { appReducers } from '../../../../store/reducers.module';
-import { AppStoreModule } from '../../../../store/store.module';
-import { generateTestApplicationServiceProvider } from '../../../../test-framework/application-service-helper';
-import { generateTestEntityServiceProvider } from '../../../../test-framework/entity-service.helper';
-import { getInitialTestStoreState } from '../../../../test-framework/store-test-helper';
+import { CoreModule } from '../../../../../../core/core.module';
+import { MDAppModule } from '../../../../../../core/md.module';
+import { LogViewerComponent } from '../../../../../../shared/components/log-viewer/log-viewer.component';
+import { ApplicationSchema, GetApplication } from '../../../../../../store/actions/application.actions';
+import { appReducers } from '../../../../../../store/reducers.module';
+import { AppStoreModule } from '../../../../../../store/store.module';
+import { generateTestApplicationServiceProvider } from '../../../../../../test-framework/application-service-helper';
+import { generateTestEntityServiceProvider } from '../../../../../../test-framework/entity-service.helper';
+import { getInitialTestStoreState } from '../../../../../../test-framework/store-test-helper';
 import { ApplicationEnvVarsService } from '../build-tab/application-env-vars.service';
-import { ApplicationStateService } from '../../../../shared/components/application-state/application-state.service';
+import { ApplicationStateService } from '../../../../../../shared/components/application-state/application-state.service';
 import { LogStreamTabComponent } from './log-stream-tab.component';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
