+++ src/app/features/applications/application/application-tabs-base/tabs/variables-tab/variables-tab.component.spec.ts
@@ -2,15 +2,15 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { StoreModule } from '@ngrx/store';
 
-import { CoreModule } from '../../../../core/core.module';
-import { MDAppModule } from '../../../../core/md.module';
-import { SharedModule } from '../../../../shared/shared.module';
-import { appReducers } from '../../../../store/reducers.module';
-import { ApplicationServiceMock } from '../../../../test-framework/application-service-helper';
-import { getInitialTestStoreState } from '../../../../test-framework/store-test-helper';
-import { ApplicationService } from '../../application.service';
+import { CoreModule } from '../../../../../../core/core.module';
+import { MDAppModule } from '../../../../../../core/md.module';
+import { SharedModule } from '../../../../../../shared/shared.module';
+import { appReducers } from '../../../../../../store/reducers.module';
+import { ApplicationServiceMock } from '../../../../../../test-framework/application-service-helper';
+import { getInitialTestStoreState } from '../../../../../../test-framework/store-test-helper';
+import { ApplicationService } from '../../../../application.service';
 import { ApplicationEnvVarsService } from '../build-tab/application-env-vars.service';
-import { ApplicationStateService } from '../../../../shared/components/application-state/application-state.service';
+import { ApplicationStateService } from '../../../../../../shared/components/application-state/application-state.service';
 import { VariablesTabComponent } from './variables-tab.component';
 
 describe('VariablesTabComponent', () => {
